Melhorias entre mapeamento de entidades
- Criada classe abstrata que recebe o id gerado automaticamente;
- Alteração do id das entidades e em alguns parâmetros de métodos para receber Long.

+++ src/main/java/marhlonkorb/github/io/gerenciadorestacionamento/models/repositories/PessoaRepository.java
@@ -1,9 +1,9 @@
 package marhlonkorb.github.io.gerenciadorestacionamento.models.repositories;
 
-import marhlonkorb.github.io.gerenciadorestacionamento.models.entities.Pessoa;
+import marhlonkorb.github.io.gerenciadorestacionamento.models.entities.pessoa.Pessoa;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PessoaRepository extends JpaRepository<Pessoa, Integer> {
+public interface PessoaRepository extends JpaRepository<Pessoa, Long> {
 }
