+++ src/main/java/marhlonkorb/github/io/gerenciadorestacionamento/models/repositories/VeiculoRepository.java
@@ -1,10 +1,10 @@
 package marhlonkorb.github.io.gerenciadorestacionamento.models.repositories;
 
-import marhlonkorb.github.io.gerenciadorestacionamento.models.entities.Veiculo;
+import marhlonkorb.github.io.gerenciadorestacionamento.models.entities.veiculo.Veiculo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VeiculoRepository extends JpaRepository<Veiculo, Integer> {
+public interface VeiculoRepository extends JpaRepository<Veiculo, Long> {
 
 }
